+ | front sensor added for the eyes

+++ src/threadx/SennaTech/Core/Inc/display/ssd1305.h
@@ -12,13 +12,13 @@ extern SPI_HandleTypeDef hspi3;
 
 void ssd1305_init();
 void ssd1305_clear(void);
-void ssd1305_fill(UINT value);
+void ssd1305_fill(uint8_t value);
 void ssd1305_update(void);
-void ssd1305_draw_pixel(UINT x, UINT y, UINT color);
-void ssd1305_draw_line(UINT x, UINT y,
-        UINT len, UINT color);
-void ssd1305_fill_ellipse(UINT cx, UINT cy,
-        UINT rx, UINT ry, UINT color);
+void ssd1305_draw_pixel(uint8_t x, uint8_t y, uint8_t color);
+void ssd1305_draw_line(uint8_t x, uint8_t y,
+        uint8_t len, uint8_t color);
+void ssd1305_fill_ellipse(uint8_t cx, uint8_t cy,
+        uint8_t rx, uint8_t ry, uint8_t color);
 
 // eye designs
 void ssd1305_draw_happy_eyes(UINT frame);
